requirements.txt
Flask
Flask-Cors
Flask-SQLAlchemy
requests
gunicorn
Flask-Mail
email-validator
itsdangerous
